enemy health scales with difficulty

+++ Common/src/main/java/org/sdu/sem4/g7/common/services/IDifficultyService.java
@@ -8,6 +8,6 @@ public interface IDifficultyService {
 
     public void setDifficulty(DifficultyEnum difficulty);
 
-    public int getMultiplier();
-    
+    public float getMultiplier();
+
 }
